Asymmetric adjacent matrix - Topographic least cost paths and distances

Calculates topographic least cost distances and paths with an asymmetric adjacent matrix

Usage

adjtopolcp(
  topography,
  costSurface,
  slopeRdCond,
  pts,
  directions = 4,
  paths = TRUE,
  zweight = 1,
  advancedloggingparameters = loggingparameters()
)

Arguments

topography

Digital terrain model (DTM) of the inventoried plot (LiDAR or SRTM) (DTMParacou) (RasterLayer with a crs in UTM)

costSurface

A RasterLayer for the conductance (inverse of resistance) values for each cell.

slopeRdCond

A categorical transition layer of longitudinal and transversal slope.

pts

A SpatialPointsDataFrame object or two-column matrix with xy coordinates for the geographic points from which to calculate pairwise distances and paths.

directions

numeric (default = 4). The number of directions for movement between cells, either 4 or 8.

paths

logical. Default is FALSE, in which case only topographic distances are calculated. If TRUE, topographic paths are also identified.

zweight

numeric (default = 1). The weight to be applied to the elevation (z) distances relative to the horizontal (xy) distances.

advancedloggingparameters

Other parameters of the logging simulator loggingparameters (list)

Value

Matrix of topographic distances (if paths = FALSE), or a list containing a matrix of topographic distances and the topographic paths as an object of class SpatialLines (if paths = TRUE).

Details

The values of the raster for costSurface should be conductance values rather than resistance values. These can be calculating by taking the inverse of resistance values.