Compute the exploitable fuel wood volume

Compute the exploitable fuel wood volume

Usage

exploitablefuelwoodvolume(
  inventory,
  scenario,
  fuel = NULL,
  TimberLoggedVolume,
  NoHollowTimberLoggedVolume,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory

Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

scenario

Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the vignette)

fuel

Fuel wood exploitation: "0": no exploitation "1": exploitation of damage and unused part of logged trees "2": exploitation of hollow trees, damage and and unused part of the log

TimberLoggedVolume

All the logged volume (in m3) (numeric)

NoHollowTimberLoggedVolume

The healthy logged volume (in m3) (numeric)

advancedloggingparameters

Other parameters of the logging simulator loggingparameters (list)

Value

A list with the damage volume, and the fuel wood volume when fuel wood exploitation is chosen.

Details

The exploitable part in fuel wood in:

the healthy trees exploited for timber: their unused part ('Purge') (Default = 0.14 m3 of purge/m3 of volume of timber harvested).

the hollow trees: 1/3 (default value of 'TreeHollowPartForFuel') of the log if fuel="2", 0 if fuel = "0" or "1".

the damage trees (trails, secondary windfall): their log volume.

Examples

data(Paracou6_2016)
data(DTMParacou)
data(MainTrails)
data(HarvestableAreaOutputsCable)

pol1 <- list(matrix(c(286503, 583134,
                      286503, 583240,
                      286507, 583240,
                      286507, 583134,
                      286503, 583134) # the return
                   ,ncol=2, byrow=TRUE))
pol2 <- list(matrix(c(286650, 583134,
                      286650, 583240,
                      286654, 583240,
                      286654, 583134,
                      286650, 583134) # the return
                   ,ncol=2, byrow=TRUE))

PolList = list(pol1,pol2) #list of lists of numeric matrices
ScndTrail <- sf::st_multipolygon(PolList)

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- suppressMessages(treeselection(inventory,
topography = DTMParacou,
speciescriteria = SpeciesCriteria,
scenario = "manual", objective = 10, fuel = "2", diversification = TRUE,
winching = "0", specieslax = FALSE, objectivelax = TRUE,
plotslope = HarvestableAreaOutputsCable$PlotSlope,
harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
advancedloggingparameters = loggingparameters())$inventory)

if (!("DeathCause" %in% names(inventory))){
  inventory <- inventory %>%
    add_column(DeathCause = NA) # if "DeathCause" column doesnt exist create it
}
inventory$DeathCause[1] <- "maintrail"
inventory$DeathCause[2] <- "2ndtrail"
inventory$DeathCause[3] <- "treefall2nd"
inventory$DeathCause[4] <- "landing"

TimberV <- timberharvestedvolume(inventory, scenario = "manual", fuel = "2",
advancedloggingparameters = loggingparameters())

TimberLoggedVolume <- TimberV$TimberLoggedVolume
NoHollowTimberLoggedVolume <- TimberV$NoHollowTimberLoggedVolume

exploitablefuelwoodvolume(inventory, scenario = "manual", fuel = "2",
TimberLoggedVolume = TimberLoggedVolume, NoHollowTimberLoggedVolume = NoHollowTimberLoggedVolume,
advancedloggingparameters = loggingparameters())
#> $DamageVolume
#> [1] 33.1903
#> 
#> $FuelVolume
#> [1] 42.97991
#>