Future & reserve trees designation

designates which trees will be harvested in future seasons: "future trees", and which should be protected as seed trees: "reserve trees".

Usage

futurereserve(
  inventory,
  speciescriteria,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory

Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

speciescriteria

Table of species exploitability criteria : species names, economic interest level, minimum and maximum felling diameter, in the same format as SpeciesCriteria (data.frame)

advancedloggingparameters

Other parameters of the logging simulator loggingparameters (list)

Value

Input inventory with selected future and reserve trees (LoggingStatus = "future"/"reserve"), for your logging criteria (SpeciesCriteria)

Details

Future trees are all trees satisfying the following conditions:

species of 1st economic rank

DBH between 35cm ('FutureTreesMinDiameter') and the species MinFD or UpMinFD if it has been raised for its species.

in the harvestable zones

Reserve trees are randomly chosen among future trees so that the number of reserve trees is equal to the number of harvested trees.

See also

Examples

data(Paracou6_2016)
data(ParamCrownDiameterAllometry)
data(ForestZoneVolumeParametersTable) # The volume parameters data in the global environment
data(SpeciesCriteria)
data(DTMParacou)
data(MainTrails)
data(HarvestableAreaOutputsCable)

scenario <- "RIL2"

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- commercialcriteriajoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory,
 topography = DTMParacou,
 diversification = TRUE,
 specieslax = FALSE,
 plotslope = HarvestableAreaOutputsCable$PlotSlope,
 maintrails = MainTrails,
 harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
 scenario = scenario,
 advancedloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

inventory <- selected(inventory, scenario = "manual", fuel = "0",
diversification = TRUE, specieslax = FALSE, objectivelax = TRUE,
topography = DTMParacou,
advancedloggingparameters = loggingparameters(), VO = 125,
 HVinit = HVinit)$inventory
#> As the harvestable volume (= 182.4m^3)
#>                 was higher (by 57.4m^3) than the objective volume,
#>                 the Minimum Falling Diameters (MinFD) of the 1st and 2nd economic ranks species were increased to UpMinFD.
#>                 Some trees with DBH lower than the UpMinFD were however selected to ensure that the objective
#>                 volume was reached.

futurereserve(inventory, SpeciesCriteria)
#> # A tibble: 3,525 × 57
#>    Forest  Plot  PlotArea SubPlot idTree Protocole Xfield Yfield UTMZone   Lat
#>    <chr>   <chr>    <dbl>   <int>  <int> <chr>      <dbl>  <dbl>   <int> <dbl>
#>  1 Paracou 6         6.25       4 103939 Guyafor    128.     34       22  5.27
#>  2 Paracou 6         6.25       2 102086 Guyafor    176.    182.      22  5.27
#>  3 Paracou 6         6.25       1 101436 Guyafor    108.    170       22  5.27
#>  4 Paracou 6         6.25       1 101413 Guyafor    106.    140.      22  5.27
#>  5 Paracou 6         6.25       2 102507 Guyafor    234.    202.      22  5.27
#>  6 Paracou 6         6.25       2 102520 Guyafor    241     181       22  5.27
#>  7 Paracou 6         6.25       2 102238 Guyafor    193     146       22  5.27
#>  8 Paracou 6         6.25       2 101998 Guyafor    172.    182.      22  5.27
#>  9 Paracou 6         6.25       4 104224 Guyafor    182.     40       22  5.27
#> 10 Paracou 6         6.25       3 103195 Guyafor     58.5    91       22  5.27
#> # ℹ 3,515 more rows
#> # ℹ 47 more variables: Lon <dbl>, Family.x <chr>, ScientificName <chr>,
#> #   Genus <chr>, Species <chr>, BotaSource <chr>, BotaCertainty <dbl>,
#> #   VernName <chr>, CensusYear <int>, CensusDateCertainty <lgl>,
#> #   CodeAlive <lgl>, MeasCode <int>, Circ <dbl>, CircCorr <dbl>,
#> #   CorrCode <chr>, DBH <dbl>, TreesIn <chr>, Xutm <dbl>, Yutm <dbl>,
#> #   TreeHarvestableVolume <dbl>, TrunkHeight <dbl>, TreeHeight <dbl>, …