Get mask

Build a mask with the intersection of the topography and the spatial inventory.

Usage

getmask(inventory, topography)

Arguments

inventory

Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

topography

Digital terrain model (DTM) of the inventoried plot (LiDAR, 1m resolution) (RasterLayer with a crs in UTM) (See DTMParacou) We advise you to generate your raster with Qgis rather than with the 'raster' package on R.

Value

The inventoried plot mask (SpatialPolygonsDataFrame with a crs in UTM)

Examples

data(Paracou6_2016)
data(DTMParacou)
mask <- getmask(Paracou6_2016, DTMParacou)