Compute the harvestable fuel wood biomass and the logging residual biomass

Computes the harvestable fuel wood biomass in healthy trees exploited for timber (their unused part), in the hollow trees and in the damage trees (caused by trails, secondary windfall). Computes also the the unused degraded tree biomass.

Usage

harvestablefuelwood(
  inventory,
  scenario,
  fuel = NULL,
  TimberLoggedVolume,
  NoHollowTimberLoggedVolume,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory

Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

scenario

Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the vignette)

fuel

Fuel wood exploitation: "0": no exploitation "1": exploitation of damage and unused part of logged trees "2": exploitation of hollow trees, damage and and unused part of the log

TimberLoggedVolume

All the logged volume (in m3) (numeric)

NoHollowTimberLoggedVolume

The healthy logged volume (in m3) (numeric)

advancedloggingparameters

Other parameters of the logging simulator loggingparameters (list)

Value

A list with the logging residual biomass (ton), the fuel wood biomass (ton) when fuel wood exploitation is chosen and the inventory with the added columns:

TimberLoggedBiomass (ton): The timber biomass logged by tree

LogBiomass (ton): the biomass of the log (equivalent to the 'TreeHarvestableVolume' but in biomass)

PurgeVolume (m3): the volume of the purge of each harvestable tree

PurgeBiomass (ton): the biomass of the purge of each harvestable tree

CrownBiomass (ton): the biomass of the tree crown

FuelWoodBiomass (ton): the fuel wood biomass harvestable by tree

LoggingResidualBiomass (ton): the unused degraded tree biomass.

Details

The harvestable part in fuel wood in:

The healthy trees exploited for timber: 2/3 of their crown (default value of 'CrownPartForFuel') + the unused part of the trunk ('Purge') (Default = 0.14 m3 of purge/m3 of volume of timber harvested).

The hollow trees if logged: 1/3 (default value of 'TreeHollowPartForFuel') of the log + 2/3 of their crown (default value of 'CrownPartForFuel')

The damage trees (trails, secondary windfall): their trunk

Examples

data(Paracou6_2016)
data(DTMParacou)
data(PlotMask)
data(ForestZoneVolumeParametersTable)
data(MainTrails)
data(HarvestableAreaOutputsCable)

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
                        volumeparameters = ForestZoneVolumeParametersTable)

inventory <- suppressMessages(treeselection(
  inventory,
  topography = DTMParacou,
  speciescriteria = SpeciesCriteria,
  scenario = "manual", objective = 10, fuel = "2", diversification = TRUE,
  winching = "0", specieslax = FALSE, objectivelax = TRUE,
  harvestablearea = HarvestableAreaOutputsCable$HarvestableArea,
  plotslope = HarvestableAreaOutputsCable$PlotSlope,
  maintrails = MainTrails,
  harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
  advancedloggingparameters = loggingparameters())$inventory)

if (!("DeathCause" %in% names(inventory))){ inventory <- inventory %>%
  tibble::add_column(DeathCause = NA) # if "DeathCause" column doesnt exist create it
}

inventory <- inventory %>%
  dplyr ::mutate(DeathCause = ifelse(is.na(DeathCause) &
                                       Selected == "1" & ProbedHollow == "0",
                                     "cut", DeathCause)) %>%
  dplyr::mutate(DeathCause = ifelse(is.na(DeathCause) &
                                      Selected == "1" & ProbedHollow == "1",
                                    "hollowfuel", DeathCause))

inventory[is.na(inventory$DeathCause), ]$DeathCause[1] <- "maintrail"
inventory[is.na(inventory$DeathCause), ]$DeathCause[2] <- "2ndtrail"
inventory[is.na(inventory$DeathCause), ]$DeathCause[3] <- "treefall2nd"
inventory[is.na(inventory$DeathCause), ]$DeathCause[4] <- "landing"

TimberV <- timberharvestedvolume(inventory, scenario = "manual", fuel = "2",
                                 advancedloggingparameters = loggingparameters())

inventory <- TimberV$inventory
TimberLoggedVolume <- TimberV$TimberLoggedVolume
NoHollowTimberLoggedVolume <- TimberV$NoHollowTimberLoggedVolume

Rslt <- harvestablefuelwood(inventory,
                            scenario = "manual", fuel = "2",
                            TimberLoggedVolume = TimberLoggedVolume,
                            NoHollowTimberLoggedVolume = NoHollowTimberLoggedVolume,
                            advancedloggingparameters = loggingparameters())