Advanced parameters of the logging simulator

Advanced parameters of the logging simulator

Usage

loggingparameters(
  MinDBHValue = 10,
  MaxTrailCenterlineSlope = 22,
  MaxTrailCrossSlope = 4,
  GrappleMaxslope = 20,
  CableTreesMaxSlope = 35,
  PlateauMaxSlope = 5,
  SlopeDistance = 3L,
  WaterSourcesBufferZone = 30,
  WaterSourcesRelativeHeight = 2,
  MinMainTrailWidth = 5,
  MaxMainTrailWidth = 6,
  ScndTrailWidth = 4,
  BigTrees = 50,
  ResamplDistDTM = 5L,
  SmoothingFact = 10,
  CableLength = 40,
  GrappleLength = 6,
  IsolateTreeMinDistance = 100,
  FutureTreesMinDiameter = 35,
  TreefallSuccessProportion = 0.6,
  MinTreefallOrientation = 30,
  MaxTreefallOrientation = 45,
  TreeHollowPartForFuel = 1/3,
  CrownPartForFuel = 2/3,
  Purge = 0.14,
  MaxTrailDensity = 200,
  MaxLandingArea = 1500,
  CostMatrix = list(list(list(Slope = 3, Cost = 3), list(Slope = 5, Cost = 5), list(Slope
    = 12, Cost = 20), list(Slope = 20, Cost = 60), list(Slope = 35, Cost = 1000),
    list(Slope = Inf, Cost = Inf)), list(list(CostType = "Initial", CostValue = 1000),
    list(CostType = "Access", CostValue = Inf), list(CostType = "BigTrees", CostValue =
    500), list(CostType = "Reserves", CostValue = 500), list(CostType = "Futures",
    CostValue = 50), list(CostType = "MainTrails", CostValue = 1e-04), list(CostType =
    "SecondTrails", CostValue = 0.1))),
  TreeHarvestableVolumeAllometry = function(DBH, aCoef, bCoef) aCoef + bCoef *
    (DBH/100)^2,
  TrunkHeightAllometry = function(DBH, TreeHarvestableVolume) TreeHarvestableVolume/(pi *
    (((DBH/100)/2)^2)),
  TreeHeightAllometry = function(DBH) exp(0.07359191 + 1.34241216 * log(DBH) +
    -0.12282344 * log(DBH)^2),
  CrownDiameterAllometry = function(DBH, TreeHeight, alpha, beta) exp(((log(DBH) - alpha
    - rnorm(length(DBH), 0, 0.0295966977))/beta))/TreeHeight,
  RottenModel = function(DBH) 1/(1 + exp(-(-5.151 + 0.042 * DBH))),
  VisiblyDefectModel = function(LogDBH) 1/(1 + exp(-(-3.392 + 0.357 * LogDBH))),
  Treefall2ndDeathModel = function(DBH) 1/(1 + exp(-(-0.47323 + -0.02564 * DBH)))
)

Arguments

MinDBHValue

Minimum DBH for inclusion in the forest inventory. Default = 10, in cm (double)

MaxTrailCenterlineSlope

Maximum trail centerline slope. Default = 22, in % (double)

MaxTrailCrossSlope

Maximum trail cross slope. Default = 4, in % (double)

GrappleMaxslope

Maximum slope accessible by the grapple. Default = 20, in % (double)

CableTreesMaxSlope

Maximum slope around the tree to access it with cable. Default = 35, in % (double)

PlateauMaxSlope

Maximum slope to define an area as a plateau. Default = 5, in % (double)

SlopeDistance

Distance over which the slope is calculated. Default = 3, in m (3m each side) (integer)

WaterSourcesBufferZone

Buffer zone based on relative horizontal distance to the nearest water source. Default = 30, in m (double)

WaterSourcesRelativeHeight

Buffer zone based on relative elevation to the nearest water source. Default = 2, in m (double)

MinMainTrailWidth

Minimum main trail width. Default = 5, in m (double)

MaxMainTrailWidth

Maximum main trail width. Default = 6, in m (double)

ScndTrailWidth

Secondary trail width. Default = 4, in m (double)

BigTrees

Minimum DBH of trees to be avoided by trails. Default = 50, in cm (double)

ResamplDistDTM

Distance of DTM resampling to erase microtopographic variation. Default = 5, in m (integer).

SmoothingFact

Secondary trails smoothing factor. Default = 10 (unitless) (double)

CableLength

Cable length. Default = 40, in m (double)

GrappleLength

Grapple length. Default = 6, in m (double)

IsolateTreeMinDistance

Minimum distance to consider a tree "isolated" from other trees of its species, in the aggregative species case (SpeciesCriteria, 'Aggregative' column). Default = 100, in m (double)

FutureTreesMinDiameter

Future trees minimum diameter. Default = 35, in cm (future trees are only commercial species of the 1st economic level) (double)

TreefallSuccessProportion

Proportion of successful directional felling events. Default = 0.6 (double)

MinTreefallOrientation

Minimum orientation of the tree fall to the trail. Default = 30, in degree (double)

MaxTreefallOrientation

Maximum orientation of the tree fall to the trail. Default = 45, in degree (double)

TreeHollowPartForFuel

Proportion of hollow trees used as fuel wood. Default = 1/3 (double)

CrownPartForFuel

Proportion of the tree crown biomass used as fuel wood. Default = 2/3 (double) (Branches diameter >= 5 cm) (Eleotério et al. 2019)

Purge

Part of the harvested log not used for timber, can be used for fuel wood. Default = 0.14, in m3 of purge/m3 of volume of timber harvested. (double)

MaxTrailDensity

Maximum trail density. Default = 200, in m/ha (double) (has no impact on the simulation. A message will be sent to inform if this threshold has been exceeded)

MaxLandingArea

Maximum landing area. Default = 1500) in m2 (double) (has no impact on the simulation. A message will be sent to inform if this threshold has been exceeded)

CostMatrix

Cost matrix for optimized trail layout (list of 2 lists). Gives an increasing cost according to a slope gradient (1st sub-list), and different costs on certain cases (2nd sub-list):

"Initial" (default = 1000)

"Access" (default = Inf)

"BigTrees" (default = 500)

"Reserves" (default = 500)

"Futures" (default = 50)

"MainTrails" (default = 1E-4)

"SecondTrails" (default = 0.1)

TreeHarvestableVolumeAllometry

By default, allometry of tree harvestable volume, French Guiana ONF formula: aCoef + bCoef * (DBH/100)^2. With aCoef and bCoef depending on the forest location, stored in ForestZoneVolumeParametersTable, DBH in cm. (function)

TrunkHeightAllometry

Allometry of trunk height, based on the cylinder volume formula: CylinderVolume = pi ((DBH/100)/2)^2 * H, with the height (H) in m and the DBH in cm (function)

TreeHeightAllometry

By default, allometry parameters estimated from Guyanese data with the BIOMASS package: ln(H) = 0.07359191 + 1.34241216 * ln(DBH) -0.12282344 * ln(DBH)^2, with the height (H) in m and the DBH in cm (function)

CrownDiameterAllometry

ln(DBH) = 𝜶 +𝜷 ln(H*CD) + 𝜺, with 𝜺~N(0,σ^2) and mean σ^2 = 0.0295966977 with the crown diameter (CD), the tree height (H) in m, and the DBH in cm. (Aubry-Kientz et al.2019)(function)

RottenModel

Estimates the tree probability of being probed hollow (default: 1 / (1 + exp(-(-5.151 + 0.042 * DBH))) with DBH in cm (developed by S.Schmitt)) (function)

VisiblyDefectModel

Estimates the commercial tree probability to have visible defects. Default: 1 / (1 + exp(-(-3.392 + 0.357 * ln(DBH)))) with DBH in cm (developed by V.Badouard) (function)

Treefall2ndDeathModel

Estimates the probability of a tree dying when it is in the area disturbed by the felling of a tree, according to the DBH of the tree whose probability of dying is estimated. Default: 1 / (1 + exp(-(-0.47323 + -0.02564 * DBH))) with DBH in cm (developed by M.Rojat) (function)

Value

A named list of 35 objects.

References

Aubry-Kientz, Mélaine, et al. "A comparative assessment of the performance of individual tree crowns delineation algorithms from ALS data in tropical forests." Remote Sensing 11.9 (2019): 1086. Eleotério, Jackson Roberto, et al. "Aboveground biomass quantification and tree-level prediction models for the Brazilian subtropical Atlantic Forest." Southern Forests: a Journal of Forest Science 81.3 (2019): 261-271.

Examples

loggingparameters(MinDBHValue = 5)
#> $MinDBHValue
#> [1] 5
#> 
#> $MaxTrailCenterlineSlope
#> [1] 22
#> 
#> $MaxTrailCrossSlope
#> [1] 4
#> 
#> $GrappleMaxslope
#> [1] 20
#> 
#> $CableTreesMaxSlope
#> [1] 35
#> 
#> $PlateauMaxSlope
#> [1] 5
#> 
#> $SlopeDistance
#> [1] 3
#> 
#> $WaterSourcesBufferZone
#> [1] 30
#> 
#> $WaterSourcesRelativeHeight
#> [1] 2
#> 
#> $MinMainTrailWidth
#> [1] 5
#> 
#> $MaxMainTrailWidth
#> [1] 6
#> 
#> $ScndTrailWidth
#> [1] 4
#> 
#> $BigTrees
#> [1] 50
#> 
#> $ResamplDistDTM
#> [1] 5
#> 
#> $SmoothingFact
#> [1] 10
#> 
#> $CableLength
#> [1] 40
#> 
#> $GrappleLength
#> [1] 6
#> 
#> $IsolateTreeMinDistance
#> [1] 100
#> 
#> $FutureTreesMinDiameter
#> [1] 35
#> 
#> $TreefallSuccessProportion
#> [1] 0.6
#> 
#> $MinTreefallOrientation
#> [1] 30
#> 
#> $MaxTreefallOrientation
#> [1] 45
#> 
#> $TreeHollowPartForFuel
#> [1] 0.3333333
#> 
#> $CrownPartForFuel
#> [1] 0.6666667
#> 
#> $Purge
#> [1] 0.14
#> 
#> $MaxTrailDensity
#> [1] 200
#> 
#> $MaxLandingArea
#> [1] 1500
#> 
#> $CostMatrix
#> $CostMatrix[[1]]
#> $CostMatrix[[1]][[1]]
#> $CostMatrix[[1]][[1]]$Slope
#> [1] 3
#> 
#> $CostMatrix[[1]][[1]]$Cost
#> [1] 3
#> 
#> 
#> $CostMatrix[[1]][[2]]
#> $CostMatrix[[1]][[2]]$Slope
#> [1] 5
#> 
#> $CostMatrix[[1]][[2]]$Cost
#> [1] 5
#> 
#> 
#> $CostMatrix[[1]][[3]]
#> $CostMatrix[[1]][[3]]$Slope
#> [1] 12
#> 
#> $CostMatrix[[1]][[3]]$Cost
#> [1] 20
#> 
#> 
#> $CostMatrix[[1]][[4]]
#> $CostMatrix[[1]][[4]]$Slope
#> [1] 20
#> 
#> $CostMatrix[[1]][[4]]$Cost
#> [1] 60
#> 
#> 
#> $CostMatrix[[1]][[5]]
#> $CostMatrix[[1]][[5]]$Slope
#> [1] 35
#> 
#> $CostMatrix[[1]][[5]]$Cost
#> [1] 1000
#> 
#> 
#> $CostMatrix[[1]][[6]]
#> $CostMatrix[[1]][[6]]$Slope
#> [1] Inf
#> 
#> $CostMatrix[[1]][[6]]$Cost
#> [1] Inf
#> 
#> 
#> 
#> $CostMatrix[[2]]
#> $CostMatrix[[2]][[1]]
#> $CostMatrix[[2]][[1]]$CostType
#> [1] "Initial"
#> 
#> $CostMatrix[[2]][[1]]$CostValue
#> [1] 1000
#> 
#> 
#> $CostMatrix[[2]][[2]]
#> $CostMatrix[[2]][[2]]$CostType
#> [1] "Access"
#> 
#> $CostMatrix[[2]][[2]]$CostValue
#> [1] Inf
#> 
#> 
#> $CostMatrix[[2]][[3]]
#> $CostMatrix[[2]][[3]]$CostType
#> [1] "BigTrees"
#> 
#> $CostMatrix[[2]][[3]]$CostValue
#> [1] 500
#> 
#> 
#> $CostMatrix[[2]][[4]]
#> $CostMatrix[[2]][[4]]$CostType
#> [1] "Reserves"
#> 
#> $CostMatrix[[2]][[4]]$CostValue
#> [1] 500
#> 
#> 
#> $CostMatrix[[2]][[5]]
#> $CostMatrix[[2]][[5]]$CostType
#> [1] "Futures"
#> 
#> $CostMatrix[[2]][[5]]$CostValue
#> [1] 50
#> 
#> 
#> $CostMatrix[[2]][[6]]
#> $CostMatrix[[2]][[6]]$CostType
#> [1] "MainTrails"
#> 
#> $CostMatrix[[2]][[6]]$CostValue
#> [1] 1e-04
#> 
#> 
#> $CostMatrix[[2]][[7]]
#> $CostMatrix[[2]][[7]]$CostType
#> [1] "SecondTrails"
#> 
#> $CostMatrix[[2]][[7]]$CostValue
#> [1] 0.1
#> 
#> 
#> 
#> 
#> $TreeHarvestableVolumeAllometry
#> function (DBH, aCoef, bCoef) 
#> aCoef + bCoef * (DBH/100)^2
#> <environment: 0x55d7bfa69378>
#> 
#> $TrunkHeightAllometry
#> function (DBH, TreeHarvestableVolume) 
#> TreeHarvestableVolume/(pi * (((DBH/100)/2)^2))
#> <environment: 0x55d7bfa69378>
#> 
#> $TreeHeightAllometry
#> function (DBH) 
#> exp(0.07359191 + 1.34241216 * log(DBH) + -0.12282344 * log(DBH)^2)
#> <environment: 0x55d7bfa69378>
#> 
#> $CrownDiameterAllometry
#> function (DBH, TreeHeight, alpha, beta) 
#> exp(((log(DBH) - alpha - rnorm(length(DBH), 0, 0.0295966977))/beta))/TreeHeight
#> <environment: 0x55d7bfa69378>
#> 
#> $RottenModel
#> function (DBH) 
#> 1/(1 + exp(-(-5.151 + 0.042 * DBH)))
#> <environment: 0x55d7bfa69378>
#> 
#> $VisiblyDefectModel
#> function (LogDBH) 
#> 1/(1 + exp(-(-3.392 + 0.357 * LogDBH)))
#> <environment: 0x55d7bfa69378>
#> 
#> $Treefall2ndDeathModel
#> function (DBH) 
#> 1/(1 + exp(-(-0.47323 + -0.02564 * DBH)))
#> <environment: 0x55d7bfa69378>
#>