loggingsimulation

This function allows to simulate a timber and wood energy logging on a forest plot. It covers: harvestable area definition, tree selection, main (only for ONF plots) and secondary trails layout, tree felling, landings implementation (only for ONF plots), timber harvested, fuel wood volume and damages quantification.

loggingsimulation(
  inventory,
  topography,
  creekdistances,
  speciescriteria,
  volumeparameters,
  scenario,
  objective = NULL,
  fuel = NULL,
  diversification = NULL,
  winching = NULL,
  directionalfelling = NULL,
  specieslax = FALSE,
  objectivelax = FALSE,
  crowndiameterparameters = ParamCrownDiameterAllometry,
  advancedloggingparameters = loggingparameters(),
  iter = 1,
  cores = 1
)

Arguments

inventory
Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

topography
Digital terrain model (DTM) of the inventoried plot (LiDAR or SRTM) (DTMParacou) (RasterLayer)

creekdistances
(RasterLayer)

speciescriteria
Table of species exploitability criteria : species names, economic interest level, minimum and maximum felling diameter, in the same format of SpeciesCriteria (2 levels of commercial species) (data.frame)

volumeparameters
Volume parameters table (in the same format of ForestZoneVolumeParametersTable) to compute the harvestable volume of each tree, depend to its geographic zone if several locations (data.frame)

scenario
Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the vignette)

objective
Objective volume (m^3/ha) (numeric)

fuel
Fuel wood exploitation: no exploitation = "0", exploitation of damage and unused part of logged trees for fuel = "1", exploitation of hollow trees, damage and and unused part of the log for fuel = "2"

diversification
Taking of other species in addition to the main commercial species (2 levels of commercial species in the SpeciesCriteria table) (logical)

winching
No cable or grapple = "0", only cable = "1", grapple + cable = "2"

directionalfelling
Directional felling = "0": only to direct the foot of the tree towards the trail "1": to direct the foot of the tree towards the trail + to avoid damage to future and reserve trees "2": to avoid damage to future and reserve trees + orientation angle to the trail

specieslax
Allow diversification if stand is too poor, = FALSE by default (logical)

objectivelax
Allow exploitation in case of non-achievement of the objective volume (if stand too poor), = FALSE by default (logical)

crowndiameterparameters
Crown diameter allometry parameters table (in the same format of ParamCrownDiameterAllometry) to compute the crown diameter of each tree, depend to its DBH (Diameter at Breast Height) and its species, genus or family. (data.frame)

advancedloggingparameters
Other parameters of the logging simulator loggingparameters (list)

iter
Number of iterations (numeric). Default = 1.

cores
Number of cores for parallelization (numeric). Default = 1.

Value

Input inventory (data.frame) with logging informations (see the outputs metadata in the vignette).

See also

Examples

if (FALSE) {
data(Paracou6_2016) # inventory
data(DTMParacou) # topography
# data() relative elevation
data(SpeciesCriteria) # species exploitability criteria
data(ForestZoneVolumeParametersTable) # volume parameters
data(ParamCrownDiameterAllometry) # parameters values of the crown diameter allometry

Rslt <- loggingsimulation(Paracou6_2016, topography = DTMParacou,
 creekdistances  = DTMParacou, speciescriteria = SpeciesCriteria,
 volumeparameters = ForestZoneVolumeParametersTable, scenario = "manual",
 objective = 20, fuel = "2", diversification = TRUE, winching = "2",
 directionalfelling = "2", specieslax = FALSE, objectivelax = TRUE,
 crowndiameterparameters = ParamCrownDiameterAllometry,
 advancedloggingparameters = loggingparameters(), iter = 1, cores = 1)
 }