Summary of the outputs of the loggingsimulation1 function

Summary of the outputs of the loggingsimulation1 function

Usage

loggingsummary1(x)

Arguments

x

Outputs of the loggingsimulation function (list)

Value

A summary of the outputs of the loggingsimulation1 function (console display):

a reminder of the inputs

the outgoing figures

a print of the output inventory Please note that all volumes in m3/ha and biomass in ton/ha are per exploitable hectare, not per plot hectare.

Examples

data(LoggingSimulationOutputs) # Outputs of one logging simulation

loggingsummary1(LoggingSimulationOutputs)
#> inventory : Paracou6_2016 
#> scenario : manual 
#> objective : 10 m3/harvestable ha
#> fuel : 2 
#> diversification : TRUE 
#> winching : 2 
#> directionalfelling : 2 
#> specieslax : FALSE 
#> objectivelax : TRUE 
#> Harvestable area : 3.75 ha
#> Objective volume : 37.5 m3
#> Initial harvestable volume : 202.4 m3,  54 m3/harvestable ha
#> Timber logged volume : 35.8 m3,  9.5 m3/harvestable ha
#> No hollow timber logged volume : 35.8 m3,  9.5 m3/harvestable ha
#> Timber extracted volume (timber volume after purge : 30.8 m3,  8.2 m3/harvestable ha
#> Fuel wood biomass : 56.5 ton,  15.1 ton/harvestable ha
#> Logging residual biomass : 5.6 ton,  1.5 ton/harvestable ha
#> Total forest biomass lost : 108.3 ton
#> Trails density : 39.6 m/ha
#> Adjusted trails density : 43.1 m/ha
#> 
#> Please note that all volumes in m3/ha are per exploitable hectare, not per plot hectare.