Selection of trees to log

Selection of trees to log among the exploitable trees, according to the objective volume and the exploitable volume on the plot.

Usage

selected(
  inventory,
  topography,
  scenario,
  fuel = NULL,
  diversification = NULL,
  VO,
  HVinit,
  specieslax = FALSE,
  objectivelax = FALSE,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory

Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

topography

Digital terrain model (DTM) of the inventoried plot (LiDAR or SRTM) (DTMParacou) (RasterLayer with a crs in UTM) We advise you to generate your raster with Qgis rather than with the 'raster' package on R.

scenario

Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the vignette)

fuel

Fuel wood exploitation: no exploitation = "0", exploitation of damage and unused part of logged trees for fuel = "1", exploitation of hollow trees, damage and and unused part of the log for fuel = "2"

diversification

Possibility to log other species in addition to the main commercial species (species with a value of 2 for commercial in the SpeciesCriteria table) (logical)

VO

Objective volume for the entire area of the plot (m3) (numeric value)

HVinit

Harvestable volume(m3) in the plot for the chosen scenario (SpeciesCriteria) (numeric value)

specieslax

Allow diversification if the stand is too poor to reach the objective volume without diversification, = FALSE by default (logical)

objectivelax

Allow exploitation in case of non-achievement of the objective volume (if stand too poor), = FALSE by default (logical)

advancedloggingparameters

Other parameters of the logging simulator loggingparameters (list)

Value

Input inventory with:

The trees selected for harvesting (Selected)

if the Minimum Felling Diameter (MinFD) of their species has been raised (Up)

The probability of a tree being probed hollow (ProbedHollowProba) and the probed hollow trees (ProbedHollow). See the outputs metadata in the vignette.

2 sets of spatial points: (HollowTrees and EnergywoodTrees)

Details

This function allows the selection of trees to be exploited among the exploitable trees. It also informs about the trees that were finally found to be hollow during the probe. If fuel = 2, the hollow trees will be harvested and are therefore included in the objective volume. If fuel = 0 or 1, the hollow trees will not be exploited, so the function looks for other trees to reach the objective volume (if possible).

If the harvestable volume is higher than the objective volume, and that diversification was not chosen, MinFD of the 1st economic rank species only is increased. If the diversification is allowed, MinFD of 1st and 2nd economic level species is increased. Then, the trees to be harvested are chosen in decreasing order of volume, until the objective volume is reached.

If the harvestable volume is lower than the objective volume, diversification can be applied if it was not already applied ('specieslax') (trees of all commercial ranks are selected in decreasing order of volume until the objective volume is reached), or harvesting can continue despite an unreached objective volume, or be abandoned ('objectivelax')

See also

Examples

data(Paracou6_2016)
data(DTMParacou)
data(PlotMask)
data(HarvestableAreaOutputsCable)
data(MainTrails)
data(ForestZoneVolumeParametersTable)

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- commercialcriteriajoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory, topography = DTMParacou,
diversification = FALSE, specieslax = TRUE,
plotslope = HarvestableAreaOutputsCable$PlotSlope,
maintrails = MainTrails,
harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
scenario = "manual", winching = "2",
advancedloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

selecInventory <- selected(inventory, topography = DTMParacou,
scenario = "manual", fuel = "2", diversification = TRUE,
VO = 20, HVinit = HVinit, specieslax = TRUE, objectivelax = TRUE,
advancedloggingparameters = loggingparameters())$inventory
#> As the harvestable volume (= 49.2m^3)
#>                 was higher (by 29.2m^3)
#>                 than the objective volume, the Minimum Falling Diameter (MinFD) of the 1st and 2nd economic ranks species
#>                   were increased to UpMinFD. The objective volume has now been reached.