Compute the timber harvested volume

Computes the timber harvested volume in healthy trees exploited for timber, and in the hollow trees if they are also harvested for fuel wood.

Usage

timberharvestedvolume(
  inventory,
  scenario,
  fuel = NULL,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory

Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

scenario

Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the vignette)

fuel

Fuel wood exploitation: no exploitation = "0", exploitation of damage and unused part of logged trees for fuel = "1", exploitation of hollow trees, damage and and unused part of the log for fuel = "2"

advancedloggingparameters

Other parameters of the logging simulator loggingparameters (list)

Value

A list with:

TimberLoggedVolume: the timber logged volume (in m3)

NoHollowTimberLoggedVolume: the timber logged volume from only healthy trees (in m3) and add in the inventory the 'TimberLoggedVolume' column to indicate the timber volume logged by tree

Details

When fuel is "2", by default, 2/3 of the log of a hollow tree will be usable as timber and 1/3 as fuel wood ('TreeHollowPartForFuel'). When fuel is not "2", hollow trees are not used for timber nor fuel wood. Dead trees from the operation (trails, secondary windfall) are never exploited as timber, but as fuel wood if chosen.

Examples

data(Paracou6_2016) # inventory
data(DTMParacou) # topography
data(MainTrails)  # MainTrails
data(HarvestableAreaOutputsCable)
data(SpeciesCriteria) # species exploitability criteria
data(ForestZoneVolumeParametersTable) # volume parameters

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- suppressMessages(treeselection(inventory,
topography = DTMParacou,
speciescriteria = SpeciesCriteria,
scenario = "manual", objective = 10, fuel = "2", diversification = TRUE,
winching = "0", specieslax = FALSE, objectivelax = TRUE,
harvestablearea = HarvestableAreaOutputsCable$HarvestableArea,
plotslope = HarvestableAreaOutputsCable$PlotSlope,
maintrails = MainTrails,
harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
advancedloggingparameters = loggingparameters())$inventory)

timberharvestedvolume(inventory, scenario = "manual", fuel = "2",
advancedloggingparameters = loggingparameters())
#> $inventory
#> # A tibble: 3,525 × 59
#>    Forest  Plot  PlotArea SubPlot idTree Protocole Xfield Yfield UTMZone   Lat
#>    <chr>   <chr>    <dbl>   <int>  <int> <chr>      <dbl>  <dbl>   <int> <dbl>
#>  1 Paracou 6         6.25       4 104455 Guyafor    215     69.5      22  5.27
#>  2 Paracou 6         6.25       4 103939 Guyafor    128.    34        22  5.27
#>  3 Paracou 6         6.25       2 102249 Guyafor    202    134.       22  5.27
#>  4 Paracou 6         6.25       2 102086 Guyafor    176.   182.       22  5.27
#>  5 Paracou 6         6.25       1 100904 Guyafor     36    170        22  5.27
#>  6 Paracou 6         6.25       3 103694 Guyafor    122    116.       22  5.27
#>  7 Paracou 6         6.25       1 101090 Guyafor     67.5  138        22  5.27
#>  8 Paracou 6         6.25       1 101193 Guyafor     75.5  208.       22  5.27
#>  9 Paracou 6         6.25       1 100833 Guyafor     25    238.       22  5.27
#> 10 Paracou 6         6.25       1 101474 Guyafor    106.   238.       22  5.27
#> # ℹ 3,515 more rows
#> # ℹ 49 more variables: Lon <dbl>, Family.x <chr>, ScientificName <chr>,
#> #   Genus <chr>, Species <chr>, BotaSource <chr>, BotaCertainty <dbl>,
#> #   VernName <chr>, CensusYear <int>, CensusDateCertainty <lgl>,
#> #   CodeAlive <lgl>, MeasCode <int>, Circ <dbl>, CircCorr <dbl>,
#> #   CorrCode <chr>, DBH <dbl>, TreesIn <chr>, Xutm <dbl>, Yutm <dbl>,
#> #   TreeHarvestableVolume <dbl>, TrunkHeight <dbl>, TreeHeight <dbl>, …
#> 
#> $TimberLoggedVolume
#> [1] 35.7794
#> 
#> $NoHollowTimberLoggedVolume
#> [1] 32.42203
#>