Logging Simulation Outputs with iteration and parallelization

Outputs of the logging simulation on the plot 6 of Paracou (French Guiana)

Usage

LoggingSimulationOutputs_iter

Format

A large list of 40 elements for each iteration, contained in a list.

inventory

Output forest inventory (data.frame)

HarvestableArea

Harvestable area in ha (numeric)

VO

Objective volume in m3 (numeric)

HVinit

Initial harvestable volume in m3 (numeric)

TimberLoggedVolume

Timber logged volume in m3 (numeric)

NoHollowTimberLoggedVolume

No hollow timber logged volume (only healthy trees) in m3 (numeric)

TimberExtractedVolume

Timber volume after purge in m3 (numeric)

FuelWoodBiomass

Harvestable fuel wood biomass in ton (numeric)

LoggingResidualBiomass

The unused degraded biomass in ton (numeric)

LostBiomass

Total lost biomass in ton (the total forest biomass lost due to logging (the sum of the AGB of all dead trees, whatever the logging cause) (numeric)

TrailsDensity

Trails density (preliminary if FWE) in m/ha (units)

AdjustTrailsDensity

Adjusted trails density if FWE in m/ha (units)

MainTrails

Main trails (sf)

HarvestablePolygons

Harvestable zones (MULTIPOLYGON with crs)

MachinePolygons

Exploitable zones accessible to machines (MULTIPOLYGON with crs)

PlotSlope

Slopes of the plot (in radians) (RasterLayer with crs)

SmoothedTrails

Smoothed secondary trails (MULTIPOLYGON with crs)

MainTrailsAccess

Random access point of main trail for each harvestable zone (POINT with crs)

TrailsIdentity

Information on sections of the trails (matrix), with LineID: , LoggedTrees: idTree of trees reached by the trails and TypeExpl: type of winching

RawSecondTrails

Non-smoothed secondary trails (SpatialLines with crs)

CostRasterAgg

The cost raster (RasterLayer)

AdjustSmoothedTrails

Adjusted smoothed secondary trails for FWE (MULTIPOLYGON with crs)

AdjustTrailsIdentity

Adjusted version of TrailsIdentity for FWE (matrix)

AdjustRawSecondTrails

Adjusted non-smoothed secondary trails for FWE (SpatialLines with crs)

HarvestableTreesPoints

Harvestable trees points (sf)

SelectedTreesPoints

Selected trees points (sf)

FutureTreesPoints

Future trees points (sf)

ReserveTreesPoints

Reserve trees points (sf)

HollowTreesPoints

Hollow trees points (sf)

EnergywoodTreesPoints

Energywood trees points (sf)

INPUTinventory

Input forest inventory (data.frame)

scenario

Input value for 'scenario' argument(character)

objective

Input value for 'objective' argument (numeric)

fuel

Input value for 'fuel' argument (character)

diversification

Input value for 'diversification' argument (logical)

winching

Input value for 'winching' argument (numeric)

directionalfelling

Input value for 'directionalfelling' argument (character)

specieslax

Input value for 'specieslax' argument (logical)

objectivelax

Input value for 'objectivelax' argument (logical)

...

Details

With the parameters: scenario : manual objective : 10 m3/ha fuel : 2 diversification : TRUE winching : 2 directionalfelling : 2 specieslax : FALSE objectivelax : TRUE

iter = 2 cores = 2