Harvestable trees identification

The function tells you the harvestable volume in the plot, and which trees are harvestable according to your harvestability criteria

harvestable(
  inventory,
  topography,
  plotslope,
  diversification,
  specieslax = FALSE,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory
Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

topography
Digital terrain model (DTM) of the inventoried plot (LiDAR or SRTM) (DTMParacou) (RasterLayer)

plotslope
Slopes (in radians) of the inventoried plot (with a neighbourhood of 8 cells) (default: PlotSlope) (RasterLayer)

diversification
Taking of other species in addition to the main commercial species (2 levels of commercial species in the SpeciesCriteria table) (logical)

specieslax
Allow diversification if stand is too poor, = FALSE by default (logical)

advancedloggingparameters
Other parameters of the logging simulator (loggingparameters) (list)

Value

input inventory with new columns:

The exploitability criteria ("DistCriteria", "Slope"(in radians), "SlopeCriteria"), and if they are validated for each of the trees ("LoggingStatus").

The probability of a tree having visible defects ("VisibleDefectProba") and the visible defect trees ("VisibleDefect").

The function returns the harvestable volume too, in the plot for these criteria.

Details

Trees will be designated as "harvestable" if they:

belonging to species of 1st economic rank or more if diversification

DBH between the MinFD and the MaxFD.

not isolated ( >100m ('IsolateTreeMinDistance' in loggingparameters)) from other individuals of the same species in the aggregative species case (SpeciesCriteria, 'Aggregative' column).

on slopes < 22% ('TreeMaxSlope'in loggingparameters)

off the main trails.

Trees with visible defects are identified ('VisiblyDefectModel' in 'advancedloggingparameters' argument) among the trees with harvestable criteria and are therefore considered 'non-harvestable'.

See also

Examples

data(Paracou6_2016)
data(DTMParacou)
data(PlotSlope)
data(SpeciesCriteria)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- commercialcriteriajoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory, topography = DTMParacou,
plotslope = PlotSlope, diversification = TRUE, specieslax = FALSE,
advancedloggingparameters = loggingparameters())

new <- harvestableOutputs$inventory