Harvestable trees identification

The function tells you the harvestable volume in the plot, and which trees are harvestable according to your harvestability criteria

Usage

harvestable(
  inventory,
  topography,
  diversification,
  specieslax = FALSE,
  harvestablepolygons,
  plotslope,
  scenario,
  winching = NULL,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory

Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

topography

Digital terrain model (DTM) of the inventoried plot (LiDAR or SRTM) (DTMParacou) (RasterLayer with a crs in UTM) We advise you to generate your raster with Qgis rather than with the 'raster' package on R.

diversification

Taking of other species in addition to the main commercial species (2 levels of commercial species in the SpeciesCriteria table) (logical)

specieslax

Allow diversification if stand is too poor, = FALSE by default (logical)

harvestablepolygons

Accessible area of the inventoried plot (default: harvestableareadefinition) (sfc_MULTIPOLYGON)

plotslope

Slopes (in radians) of the inventoried plot (with a neighbourhood of 8 cells) (default: HarvestableAreaOutputsCable) (RasterLayer with a crs in UTM)

scenario

Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the vignette)

winching

"0": no cable or grapple (trail to tree foot) "1": only cable (default = 40m) "2": grapple (default = 6m) + cable (grapple priority) If grapple + cable (winching = "2") without fuel wood (fuel = "0") recovery of the tree foot with grapple if possible (respected grapple conditions) otherwise with cable with angle to the trail. Avoidance of future/reserves if chosen.

advancedloggingparameters

Other parameters of the logging simulator (loggingparameters) (list)

Value

input inventory with new columns:

The exploitability criteria ("DistCriteria", "Slope"(in radians), "SlopeCriteria"), and if they are validated for each of the trees ("LoggingStatus").

The probability of a tree having visible defects ("VisibleDefectProba") and the visible defect trees ("VisibleDefect").

The function returns the harvestable volume too, in the plot for these criteria.

Details

Trees will be designated as "harvestable" if they:

belonging to species of 1st economic rank or more if diversification

DBH between the MinFD and the MaxFD.

not isolated ( >100m ('IsolateTreeMinDistance' in loggingparameters)) from other individuals of the same species in the aggregative species case (SpeciesCriteria, 'Aggregative' column).

on slopes < 22% ('TreeMaxSlope'in loggingparameters)

off the main trails.

Trees with visible defects are identified ('VisiblyDefectModel' in 'advancedloggingparameters' argument) among the trees with harvestable criteria and are therefore considered 'non-harvestable'.

See also

Examples

data(Paracou6_2016)
data(DTMParacou)
data(PlotMask)
data(SpeciesCriteria)
data(HarvestableAreaOutputsCable)

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- commercialcriteriajoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory, topography = DTMParacou,
diversification = TRUE, specieslax = FALSE,
plotslope = HarvestableAreaOutputsCable$PlotSlope,
harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
scenario = "manual", winching = "0",
advancedloggingparameters = loggingparameters())

new <- harvestableOutputs$inventory