Summary of the outputs of the loggingsimulation function

Summary of the outputs of the loggingsimulation function

Usage

loggingsummary(x)

Arguments

x

Outputs of the loggingsimulation function (list)

Value

A summary of the outputs of the loggingsimulation function (console display):

a reminder of the inputs

the iterations statistics (mean, standard deviation, quantiles) of the outgoing figures Please note that all volumes in m3/ha and biomass in ton/ha are per exploitable hectare, not per plot hectare.

Examples

data(LoggingSimulationOutputs_iter) # Outputs of one logging simulation (2 iterations and 2 cores)

loggingsummary(x = LoggingSimulationOutputs_iter)
#> inventory : inventory 
#> scenario : manual 
#> objective : 10 m3/harvestable ha
#> fuel : 2 
#> diversification : TRUE 
#> winching : 0 
#> directionalfelling : 2 
#> specieslax : FALSE 
#> objectivelax : TRUE 
#> Harvestable area (ha):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>    2.3    0.0    2.3    2.0    2.3    2.0    2.3 
#> 
#> Objective volume (m3):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>   23.3    0.0   23.3   23.0   23.3   23.0   23.3 
#> 
#> Initial harvestable volume (m3):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>   91.0    9.2   84.5   88.0   91.0   94.0   97.5 
#> Units: [m3/ha]
#>     Mean       Sd      Min   Q1.25%   Median   Q3.75%      Max 
#> 39.56522      Inf 36.73913 44.00000 39.56522 47.00000 42.39130 
#> 
#> Timber logged volume (m3):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>   21.3    0.0   21.3   21.0   21.3   21.0   21.3 
#> Units: [m3/ha]
#>     Mean       Sd      Min   Q1.25%   Median   Q3.75%      Max 
#>  9.26087      NaN  9.26087 10.50000  9.26087 10.50000  9.26087 
#> 
#> No hollow timber logged volume (m3):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>   17.3    0.0   17.3   17.0   17.3   17.0   17.3 
#> Units: [m3/ha]
#>     Mean       Sd      Min   Q1.25%   Median   Q3.75%      Max 
#> 7.521739      NaN 7.521739 8.500000 7.521739 8.500000 7.521739 
#> 
#> Timber extracted volume (timber volume after purge) (m3):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>   18.9    0.0   18.9   19.0   18.9   19.0   18.9 
#> Units: [m3/ha]
#>     Mean       Sd      Min   Q1.25%   Median   Q3.75%      Max 
#> 8.217391      NaN 8.217391 9.500000 8.217391 9.500000 8.217391 
#> 
#> Fuel wood biomass (ton):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>   41.8    3.9   39.0   40.0   41.8   43.0   44.6 
#> Units: [ton/ha]
#>     Mean       Sd      Min   Q1.25%   Median   Q3.75%      Max 
#> 18.17391      Inf 16.95652 20.00000 18.17391 21.50000 19.39130 
#> 
#> Logging residual biomass (ton):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>    4.1    0.0    4.1    4.0    4.1    4.0    4.1 
#> Units: [ton/ha]
#>     Mean       Sd      Min   Q1.25%   Median   Q3.75%      Max 
#> 1.782609      NaN 1.782609 2.000000 1.782609 2.000000 1.782609 
#> 
#> Total forest biomass lost (ton):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>   76.7    7.1   71.7   74.0   76.7   79.0   81.7 
#> 
#> Trails density (m/ha):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>   45.3    4.9   41.9   44.0   45.3   47.0   48.8 
#> 
#> Adjusted trails density (m/ha):
#>   Mean     Sd    Min Q1.25% Median Q3.75%    Max 
#>   45.3    4.9   41.9   44.0   45.3   47.0   48.8 
#> 
#> Please note that all volumes in m3/ha and biomass in ton/ha are per exploitable hectare, not per plot hectare.