Selection of trees to log

Selection of trees to log

selected(
  inventory,
  topography,
  scenario,
  fuel = NULL,
  diversification = NULL,
  VO,
  HVinit,
  specieslax = FALSE,
  objectivelax = FALSE,
  advancedloggingparameters = loggingparameters()
)

Arguments

inventory
Input inventory (see the inputs formats and metadata in the vignette) (data.frame)

topography
Digital terrain model (DTM) of the inventoried plot (LiDAR or SRTM) (DTMParacou) (RasterLayer)

scenario
Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the vignette)

fuel
Fuel wood exploitation: no exploitation = "0", exploitation of damage and unused part of logged trees for fuel = "1", exploitation of hollow trees, damage and and unused part of the log for fuel = "2"

diversification
TPossibility to log other species in addition to the main commercial species (species with a value of 2 for commercial in the SpeciesCriteria table) (logical)

VO
Objective volume for the entire area of the plot (numeric value)

HVinit
Harvestable volume in the plot for the chosen scenario (SpeciesCriteria) (numeric value)

specieslax
Allow diversification if the stand is too poor to reach the objective volume without diversification, = FALSE by default (logical)

objectivelax
Allow exploitation in case of non-achievement of the objective volume (if stand too poor), = FALSE by default (logical)

advancedloggingparameters
Other parameters of the logging simulator loggingparameters (list) MainTrails (multiline)

Value

Input inventory with

The trees selected for harvesting ("Selected")

if the Minimum Felling Diameter (MinFD) of their species has been raised ("Up")

The cumulative harvestable volume of harvestable trees("VolumeCumSum").

The probability of a tree being probed hollow ("ProbedHollowProba") and the probed hollow trees ("ProbedHollow"). See the outputs metadata in the vignette

2 sets of spatial points: (HollowTrees and EnergywoodTrees)

Details

This function allows the selection of trees to be exploited among the exploitable trees. It also informs about the trees that were finally found to be hollow during the probe. If fuel = 2, the hollow trees will be harvested and are therefore included in the objective volume. If fuel = 0 or 1, the hollow trees will not be exploited, so the function looks for other trees to reach the objective volume (if possible).

If the harvestable volume is higher than the objective volume, MinFD of the 1st economic rank species is increased. If this is not enough and if diversification is allowed, MinFD of 2nd economic level species is increased. Then, the trees to be harvested are chosen in decreasing order of volume, until the objective volume is reached.

If the harvestable volume is lower than the objective volume, diversification can be applied if it was not already applied ('specieslax') (trees of all commercial ranks are selected in decreasing order of volume until the objective volume is reached), or harvesting can continue despite an unreached objective volume, or be abandoned ('objectivelax')

See also

Examples

data(Paracou6_2016)
data(DTMParacou)
data(PlotSlope)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- commercialcriteriajoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory, topography = DTMParacou,
diversification = TRUE, specieslax = FALSE, plotslope = PlotSlope,
advancedloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

selecInventory <- selected(inventory, topography = DTMParacou,
scenario = "manual", fuel = "2", diversification = TRUE,
VO = 125, HVinit = HVinit, specieslax = FALSE, objectivelax = TRUE,
advancedloggingparameters = loggingparameters())$inventory

#> As the harvestable volume (= 194.1m^3)
#>                 was higher (by 69.1m^3)
#>                 than the objective volume, it was necessary to increase the Minimum Falling Diameter (MinFD)
#>                 of all species. The objective volume has now been reached.